+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Add
+- Cmap for Mutant Visualizer
+
 ## [1.1.2] - 2023-10-09
 
 ### Add